Split "Draw Card" button into two buttons, "Draw From Top" and "Draw From Bottom".
Also added Debug.Log() functionality to both Draw Card functions as well as the shuffle function that is currently called upon run. Now, when the deck is shuffled, a full list of everything in the deck is printed to the log in order. This makes it very easy to check and ensure that "Draw From Top" draws from the "top" and "Draw From Bottom" draws from the "bottom" of the deck.